Site root now contains a flow of posts (UGLY:
The only way to result the default and post layout conundrum
and to be DRY, the html code around the Liquid {{ content }}
statement has been moved to _include folders. This way it can
be iterated through with existing content.

CAVEAT: I don't know how badly it would affect page load and
google analytics...